downloader/requirements.txt
gallery-dl>=1.31.0
yt-dlp>=2026.3.0
cloudinary>=1.41.0
python-dotenv>=1.0.0
requests>=2.31.0
